be/requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.23.3
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.0.2
